+++ R/engine_stan.R
@@ -35,6 +35,12 @@ NULL
 #' * Carpenter, B., Gelman, A., Hoffman, M. D., Lee, D., Goodrich, B., Betancourt, M., ... & Riddell, A. (2017). Stan: A probabilistic programming language. Journal of statistical software, 76(1), 1-32.
 #' * Piironen, J., & Vehtari, A. (2017). Sparsity information and regularization in the horseshoe and other shrinkage priors. Electronic Journal of Statistics, 11(2), 5018-5051.
 #' @family engine
+#' @returns An [engine].
+#' @examples
+#' \dontrun{
+#' # Add Stan as an engine
+#' x <- distribution(background) |> engine_stan(iter = 1000)
+#' }
 #' @name engine_stan
 NULL
 #' @rdname engine_stan